API/requirements.txt
Django==1.11.3
django-extensions==1.8.1
djangorestframework==3.6.3
django-cors-headers==2.1.0
requests==2.10.0
ssh-import-id==5.6
system-service==0.3
